+++ src/Services/FhirResourcesService.php
@@ -12,16 +12,9 @@
 
 namespace OpenEMR\Services;
 
-use OpenEMR\FHIR\R4\FHIRDomainResource\FHIRPractitioner;
-use OpenEMR\FHIR\R4\FHIRElement\FHIRAddress;
-use OpenEMR\FHIR\R4\FHIRElement\FHIRAdministrativeGender;
-use OpenEMR\FHIR\R4\FHIRElement\FHIRHumanName;
-use OpenEMR\FHIR\R4\FHIRElement\FHIRId;
 use OpenEMR\FHIR\R4\FHIRResource\FHIRBundle;
 use OpenEMR\FHIR\R4\FHIRResource\FHIRBundle\FHIRBundleLink;
 use OpenEMR\FHIR\R4\PHPFHIRResponseParser;
-
-use OpenEMR\Services\ProviderService;
 
 class FhirResourcesService
 {
@@ -49,40 +42,6 @@ class FhirResourcesService
         }
 
         return $bundle;
-    }
-
-    public function createPractitionerResource($provider_id = '', $encode = true)
-    {
-        if (!$provider_id) {
-            return false;
-        }
-        $this->providerService = new ProviderService();
-        $data = $this->providerService->getById($provider_id);
-        $resource = new FHIRPractitioner();
-        $id = new FhirId();
-        $name = new FHIRHumanName();
-        $address = new FHIRAddress();
-        $id->setValue('' . $provider_id);
-        $name->setUse('official');
-        $name->setFamily($data['lname']);
-        $name->given = [$data['fname'], $data['mname']];
-        $address->addLine($data['street']);
-        $address->setCity($data['city']);
-        $address->setState($data['state']);
-        $address->setPostalCode($data['zip']);
-        $resource->setId($id);
-        $resource->setActive(true);
-        $gender = new FHIRAdministrativeGender();
-        $gender->setValue('unknown');
-        $resource->setGender($gender);
-        $resource->addName($name);
-        $resource->addAddress($address);
-
-        if ($encode) {
-            return json_encode($resource);
-        } else {
-            return $resource;
-        }
     }
 
     public function parseResource($rjson = '', $scheme = 'json')
